requirements.txt
streamlit
nemo_toolkit[asr]
numpy==1.26.4
Cython
setuptools
wheel